+++ examples/webgl_loader_md2_control.html
@@ -41,6 +41,7 @@
 		<script src="js/controls/OrbitControls.js"></script>
 
 		<script src='js/loaders/MD2Loader.js'></script>
+		<script src='js/MorphBlendMesh.js'></script>
 		<script src='js/MD2CharacterComplex.js'></script>
 		<script src='js/Gyroscope.js'></script>
