--- test/unit/src/extras/objects/MorphBlendMesh.js
@@ -1,6 +1,0 @@
-/**
- * @author TristanVALCKE / https://github.com/TristanVALCKE
- */
-
-//Todo
-console.warn("Todo: Unit tests of MorphBlendMesh")
