+++ test/unit/unittests_sources.html
@@ -25,7 +25,7 @@
   <script src="src/polyfills.js"></script>
   <script src="src/utils.js"></script>
 
-  
+
   <!-- /src/animation -->
   <script src="src/animation/AnimationAction.js"></script>
   <script src="src/animation/AnimationClip.js"></script>
@@ -46,7 +46,7 @@
   <script src="src/animation/tracks/StringKeyframeTrack.js"></script>
   <script src="src/animation/tracks/VectorKeyframeTrack.js"></script>
 
-  
+
   <!-- /src/audio -->
   <script src="src/audio/Audio.js"></script>
   <script src="src/audio/AudioAnalyser.js"></script>
@@ -54,7 +54,7 @@
   <script src="src/audio/AudioListener.js"></script>
   <script src="src/audio/PositionalAudio.js"></script>
 
-  
+
   <!-- /src/cameras -->
   <script src="src/cameras/Camera.js"></script>
   <script src="src/cameras/CubeCamera.js"></script>
@@ -62,7 +62,7 @@
   <script src="src/cameras/PerspectiveCamera.js"></script>
   <script src="src/cameras/StereoCamera.js"></script>
 
-  
+
   <!-- /src/core -->
   <script src="src/core/BufferAttribute.js"></script>
   <script src="src/core/BufferGeometry.js"></script>
@@ -110,7 +110,6 @@
 
   <!-- /src/extras/objects -->
   <script src="src/extras/objects/ImmediateRenderObject.js"></script>
-  <script src="src/extras/objects/MorphBlendMesh.js"></script>
 
 
   <!-- /src/geometries -->
@@ -139,7 +138,7 @@
   <script src="src/geometries/TubeGeometry.tests.js"></script>
   <script src="src/geometries/WireframeGeometry.tests.js"></script>
 
-  
+
   <!-- /src/helpers -->
   <script src="src/helpers/ArrowHelper.js"></script>
   <script src="src/helpers/AxisHelper.js"></script>
